Big commit for adhearing to AbilitySystemSetup plugin update
Plugin changes we reacted to:
	Delegates were deleted - GiveStartingAbilitiesDelegate AddStartingAttributeSetsDelegate and OnAbilitySystemSetUpPreIntialized.
	Using new AbilitySystemGrantSets

Had to remove all AbilitySpecHandles that were being replicated to the
owner for activating abilities. This cleaned up lots of code since now
all we are doing is activating by tag.
We also started following the tag member naming convention where you
just make the name the whole tag with using underscores instead of
periods. Also we wrapped the Native Gameplay Tags in a namespace and included the file in the PCH so that it is as easy as NativeGameplayTags::Some_Tag.
Also had to make new Gameplay Tags for the swap item abilities since we
are now activating by tag. This required for us to assign these tags in
the BP subclass for the swap item ability. Also made sure to have our
code enforce us to have an ability tag for all abilities.
While we were at it with the tag stuff, we also changed the naming of
some tags to be wayyyy more clear, and as a result we also had to change
the input action names for them and the ability system input enum input
names.

And all this required for us to re-assign members in our BP classes for
stuff we affected since they resetted. Also we had to create
AbilitySystemGrantSets for our characters and fill those out in their BP
since the plugin requires we use those now.