Adhere to AbilitySystemSetup base Target Actor changes.
Make MaxRange and TraceChannel members on the BulletTrace Target Actor because the base takes this out. And get rid of Super call in StartTargeting() because the base changed it now.

Have ConfirmTargetingAndContinue() use the Controller (not PlayerController) view point to be generic to AI.

And make the collision shape really small.