Realize physical animation isn't of any use for this.
Remove profile and event graph logic for the physical animation.
Also add rigidbody node in the animgraph since that seams to be the
practice when simulating physics on bodies.
Also Make end hairs collidable with all except already touching hairs.
Also fix bodies not colliding with eachother because the CharacterMeshPhysicsCollision preset was of object type of "Pawn". Change to "WorldDynamic".