Improve ActiveItem widget class and stop injecting into widgets.
Have the ActiveItem widget search externally for the active item stack and use that name for its text.

Get rid of injection into the created ActiveItem widget.

Make active item widget pointers generic to any User Widget since we
are no longer injecting into it. And delete the member that was for injecting into.