Get rid of unecessary default values in attribute sets.
These default values given to attributes make it difficult when testing to determine if attributes have been initialized properly (via a gameplay effect).

Have initialization execution calculations not copy defaults from attribute set classes.

And not do anything for the default values as they are kind of arbitrary. The game can make base classes for this if they want to.

Update AbilitySystemSetupExtras plugin to do the same.